onevault_api/requirements.txt
# OneVault API - Minimal Production Dependencies
# ===============================================

# Core FastAPI dependencies
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0

# Database
psycopg2-binary>=2.9.0,<3.0.0
asyncpg>=0.29.0,<0.30.0

# Security and validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Zero Trust Security Dependencies
sqlparse>=0.4.4,<0.5.0  # SQL query parsing for automatic tenant filtering

# Production server
gunicorn>=21.2.0,<22.0.0 